REF: make a default cli function for the differ
