+++ index.js
@@ -1,10 +1,12 @@
 var batcher = require('byte-stream')
 var through = require('through2')
 var debug = require('debug')('visualdiff')
+var prompt = require('prompt-sync')
 
 var dat2daff = require('./lib/dat2daff.js')
 
 function VisualDiff (heads, opts, cb) {
+  if (!cb) cb = VisualDiff.cli
   if (!(this instanceof VisualDiff)) return new VisualDiff(heads, opts, cb)
   /*
   strategy:
@@ -80,6 +82,60 @@ VisualDiff.prototype.decline = function () {
 VisualDiff.prototype.merge = function () {
   this.mergeStream.write()
   this.next()
+}
+
+VisualDiff.cli = function (data, visual, next) {
+  // TODO: modularize into a dat visual merge tool?
+  var self = this
+
+  var heads = data.heads
+  var tables = data.tables
+  var older = data.older // 'left' or 'right'
+
+  console.log(visual)
+
+  function repl () {
+    // TODO: change limit in repl (like git's add -p or e/edit)
+    process.stdout.write('Keep this chunk? [y,n,s,r,c,q,?] ')
+    var val = prompt()
+    if (val === 's' || val === 'skip') {
+      return next()
+    }
+    if (val === 'y' || val === 'yes') {
+      // TODO: choose 'newer' version
+      return next()
+    }
+    if (val === 'n' || val === 'no') {
+      // TODO: choose 'older' version
+      return next()
+    }
+    if (val === 'r' || val === 'rows') {
+      opts.strategy = 'rows'
+      // differ = makeDiffer(heads)
+      return
+    }
+    if (val === 'c' || val === 'cols') {
+      opts.strategy = 'cols'
+      // differ = makeDiffer(heads)
+      return
+    }
+    if (val === 'q' || val === 'quit') {
+      return process.exit()
+    }
+    else {
+      help()
+      repl()
+    }
+  }
+  repl()
+}
+
+function help () {
+  console.log('skip (s), yes (y), no (n), cols (c), rows (r), quit (q)')
+}
+
+function usage () {
+  console.log("dat-visualDiff <dat-db> [--limit <num>] [--heads <head1,head2>]")
 }
 
 module.exports = VisualDiff
